Added support for non cluster connections in the redis cache adapter
no-issue

Previously, the adapter was only built for a redis cluster connection. Meaning if you tried to use if with a redis that was a single node it would fail as it tries to find a primary node. In a single node setup there is no primary node, just the one main node. So, this update tricks the adapter into thinking it has found a pimary node by returning the whole connection (to the single node) when the constructor is note a cluster.